proyecto acabadao

+++ script.js
@@ -5,6 +5,7 @@ var palabra1 = "clan"
 var palabra2 = "pena"
 var palabra3 = "remato"
 var palabra4 = "torero"
+var cookies_aceptadas = (window.localStorage.getItem("cookies_aceptadas") == "true")
 
 function crear_tabla() {
 
@@ -21,7 +22,7 @@ function crear_tabla() {
         for (let j = 0; j < longitud; j++) {
             const celda = document.createElement("td")
             celda.className = "celda_pasatiempo"
-            celda.innerHTML = `<input type="text" maxlength="1" class="input_celda_pasatiempo" onchange="comprobar_fila(${i})">`
+            celda.innerHTML = `<input type="text" maxlength="1" class="input_celda_pasatiempo" oninput="comprobar_pasatiempo()">`
 
             fila_html.appendChild(celda)
             fila_array.push(celda)
@@ -31,7 +32,6 @@ function crear_tabla() {
     }
 
     window.localStorage.getItem("pasatiempo") && cargar_pasatiempo()
-
 }
 
 function get_palabra(n_fila){
@@ -47,8 +47,6 @@ function get_palabra(n_fila){
 function comprobar_fila(n_fila){
     const fila = tabla_pasatiempo[n_fila]
     let palabra = get_palabra(n_fila)
-
-    console.log(palabra)
 
     if (palabra.length === 4 && n_fila < 6 || palabra.length === 6 && n_fila >= 6 ) {
         switch (n_fila) {
@@ -77,11 +75,16 @@ function comprobar_fila(n_fila){
                     set_fila_erronea(fila)
                 break;
             default:
-                if (en_diccionario(palabra) && es_permutacion(n_fila))
+                if (en_diccionario(palabra) && es_permutacion(n_fila)) {
                     set_fila_correcta(fila)
-                else
+                }
+                else {
                     set_fila_erronea(fila)
+                }
                 break;
+        }
+        if(cookies_aceptadas) {
+            guardar_pasatiempo()
         }
     }else {
         set_fila_default(fila)
@@ -119,37 +122,46 @@ function es_permutacion(n_fila){
             coincidencias++;
         }
     }
-    console.log(coincidencias)
     return coincidencias >= palabra_lista.length-1;
 }
 
 function set_fila_correcta(fila){
 
     for (let i = 0; i < fila.length; i++) {
-        fila[i].className = "celda_pasatiempo fila_correcta"
+        fila[i].firstChild.className = "input_celda_pasatiempo fila_correcta"
     }
 }
 
 function set_fila_erronea(fila){
 
     for (let i = 0; i < fila.length; i++) {
-        fila[i].className = "celda_pasatiempo fila_erronea"
+        fila[i].firstChild.className = "input_celda_pasatiempo fila_erronea"
     }
 }
 
 function set_fila_default(fila){
 
     for (let i = 0; i < fila.length; i++) {
-        fila[i].className = "celda_pasatiempo"
+        fila[i].firstChild.className = "input_celda_pasatiempo"
     }
 }
 
 function mostrar_pistas(){
 
+    console.log("EYEYEY")
+    let input = document.getElementById("input_pista")
     if (pistas_restantes > 0) {
+
         pistas_restantes --;
 
-        pistas = get_palabras_pistas()
+        window.localStorage.setItem("pistas_restantes", pistas_restantes.toString())
+
+        let palabra = document.getElementById("input_pista").value
+
+        input.placeholder = `Tienes ${pistas_restantes} pistas restantes`
+        input.value = ""
+
+        pistas = get_palabras_pistas(palabra)
 
         document.getElementById("div_lista_pistas").innerHTML = ""
 
@@ -158,14 +170,18 @@ function mostrar_pistas(){
             elemento.innerHTML = pista
             document.getElementById("div_lista_pistas").appendChild(elemento)
         })
+    }else if (pistas_restantes === 0){
+        input.placeholder = "No te quedan pistas!"
     }
+    input.value = ""
 }
 
-function get_palabras_pistas() {
+function get_palabras_pistas(palabra) {
 
-    let palabra = document.getElementById("input_pista").value.split("")
+    palabra = palabra.split("")
     let pistas = []
     let regex = ""
+
 
     palabra.forEach(letra => regex += `(?=.*${letra})`)
 
@@ -188,10 +204,12 @@ function guardar_pasatiempo(){
     })
 
     window.localStorage.setItem("pasatiempo", filas.join(";"))
-    window.localStorage.setItem("pistas_restantes", pistas_restantes.toString())
 }
 
 function cargar_pasatiempo(){
+
+    pistas_restantes = window.localStorage.getItem("pistas_restantes")
+
     let tabla = window.localStorage.getItem("pasatiempo")
     let filas = tabla.split(";")
     let celdas = []
@@ -210,15 +228,84 @@ function cargar_pasatiempo(){
 }
 
 function limpiar_almacenamiento(){
-    window.localStorage.clear()
+    if (cookies_aceptadas) {
+        window.localStorage.clear()
+        window.alert("Almacenamiento borrado correctamente.")
+    }else {
+        window.alert("No hay nada que eliminar.")
+    }
 }
 
+function set_tema(nombre_tema) {
+    localStorage.setItem("tema", nombre_tema);
+    document.documentElement.className = nombre_tema;
+}
+
+function cambiar_tema() {
+    if (localStorage.getItem("tema") === "tema-oscuro") {
+        set_tema("tema-claro");
+    } else {
+        set_tema("tema-oscuro");
+    }
+}
+
+console.log(document.getElementById("input_pista"))
+
+function on_load() {
+    if (window.localStorage.getItem("pistas_restantes")){
+        pistas_restantes = window.localStorage.getItem("pistas_restantes")
+        document.getElementById("input_pista").placeholder = `Tienes ${pistas_restantes} pistas restantes`
+    }else
+        window.localStorage.setItem("pistas_restantes", pistas_restantes.toString())
+
+    document.getElementById("input_pista").addEventListener("keydown", event => {
+        if (event.key === "Enter" && document.getElementById("input_pista").value.trim() !== "") {
+            mostrar_pistas()
+        }
+    });
+}
 fetch("https://ordenalfabetix.unileon.es/aw/diccionario.txt")
-        .then(response => response.text())
-        .then (data => data.split("\n"))
-        .then (diccionario => {
-            const palabras_ejemplo = ["cria", "nací", "nace"]
-            mi_diccionario = diccionario.concat(palabras_ejemplo)
-            crear_tabla()
-        })
-        .catch( error => console.log(error));
+    .then(response => response.text())
+    .then (data => data.split("\n"))
+    .then (diccionario => {
+        const palabras_ejemplo = ["cria", "nací", "nace", "tolero"]
+        mi_diccionario = diccionario.concat(palabras_ejemplo)
+        crear_tabla()
+    })
+    .catch( error => console.log(error));
+
+function set_cookies(aceptadas){
+    cookies_aceptadas = aceptadas;
+    window.localStorage.setItem("cookies_aceptadas", aceptadas)
+    document.getElementById("cookieConsentContainer").style.visibility = "hidden"
+}
+
+(function () {
+    if (!cookies_aceptadas) {
+
+        document.body.innerHTML += "<div class=\"cookieConsentContainer\" id=\"cookieConsentContainer\" style=\"opacity: 1; display: block;\">\n" +
+            "    <div class=\"cookieTitle\">\n" +
+            "        <a>Almacenamiento local.</a>\n" +
+            "    </div>\n" +
+            "    <div class=\"cookieDesc\">\n" +
+            "        <p>¿Acepta que almacenemos datos de esta página web como la resolución parcial o el tema para su posterior recuperación?</p>\n" +
+            "    </div>\n" +
+            "        <div class=\"cookie_aceptar\">\n" +
+            "            <a onclick=\"set_cookies(true)\">Acepto</a>\n" +
+            "        </div>\n" +
+            "        <div class=\"cookie_rechazo\">\n" +
+            "            <a onclick=\"set_cookies(false)\">Rechazar</a>\n" +
+            "        </div>\n" +
+            "</div>"
+    }
+})();
+
+(function () {
+    if (localStorage.getItem('tema') === 'tema-claro') {
+        set_tema('tema-claro');
+        document.getElementById('slider').checked = true;
+    } else {
+        set_tema('tema-oscuro');
+        document.getElementById('slider').checked = false;
+    }
+})();
